requirements.txt
tavily-python==0.5.0
fastapi==0.115.2
uvicorn==0.31.1
python-dotenv==1.0.1
streamlit==1.39.0
langchain_mistralai==0.2.0 
pillow==10.4.0
srt==3.5.3
google-cloud-aiplatform==1.70.0
edge-tts==6.1.12
moviepy==1.0.3
pysrt==1.1.2
